Use geometry as submodule and fix ninja compilation errors.

+++ include/Aka/Resource/Asset.hpp
@@ -6,7 +6,7 @@ namespace aka {
 
 enum class AssetID : uint64_t 
 {
-	Invalid = -1
+	Invalid = (uint64_t)-1
 };
 
 // Adding element not at the end of this struct might invalidate library.json.
